Add a script to update the site contents.
The script takes the README.md from the master branch and uses that to
update the index.md in the gh-pages branches.

Because I have two checkouts of this repository, one for each branch, I
made the script (ab)use that. Only if it detects no other checkout, it
will switch branches.